requirements.txt
Flask
flask-cors
flask-jwt-extended
SQLAlchemy
python-dotenv
passlib[bcrypt]
bcrypt
cryptography
Pillow
qrcode
routeros-api
pytest
